+++ src/components/sync/SyncOrchestration.tsx
@@ -2,6 +2,7 @@ import { BandSync } from './BandSync';
 import { CacheVersionCheck } from './CacheVersionCheck';
 import { DuckNotificationsListener } from './DuckNotificationsListener';
 import { PushSetup } from './PushSetup';
+import { RealtimeSync } from './RealtimeSync';
 import { ReconnectSync } from './ReconnectSync';
 
 export function SyncOrchestration() {
@@ -10,6 +11,7 @@ export function SyncOrchestration() {
       <CacheVersionCheck />
       <BandSync />
       <ReconnectSync />
+      <RealtimeSync />
       <PushSetup />
       <DuckNotificationsListener />
     </>